+++ Editor/Util.cs
@@ -550,6 +550,8 @@ namespace Reallusion.Import
         {
             if (!character) return;
 
+            GameObject prefab = GetCharacterPrefab(character);
+
             if (!EditorSceneManager.SaveCurrentModifiedScenesIfUserWantsTo()) return;            
                         
             EditorSceneManager.NewScene(NewSceneSetup.EmptyScene);
@@ -559,7 +561,7 @@ namespace Reallusion.Import
             {
                 DestroyEditorChildObjects(container);
 
-                GameObject clone = PrefabUtility.InstantiatePrefab(character, container.transform) as GameObject;
+                GameObject clone = PrefabUtility.InstantiatePrefab(prefab ? prefab : character, container.transform) as GameObject;
                 if (clone)
                 {
                     Selection.activeGameObject = clone;
@@ -577,7 +579,7 @@ namespace Reallusion.Import
                 for (int i = 0; i < container.transform.childCount; i++)
                 {
                     GameObject child = container.transform.GetChild(i).gameObject;
-                    GameObject source = GetPrefabFromObject(child);
+                    GameObject source = GetSourcePrefabFromObject(child);
                     if (source == prefab)
                     {
                         Debug.Log("Keeping existing generated prefab...");
@@ -614,9 +616,21 @@ namespace Reallusion.Import
                     }
                 }                
             }
-        }                  
+        }      
+        
+        public static GameObject GetCharacterPrefab(GameObject fbx)
+        { 
+            string path = AssetDatabase.GetAssetPath(fbx);
+            if (path.iEndsWith(".prefab")) return fbx;
+            string folder = Path.GetDirectoryName(path);
+            string name = Path.GetFileNameWithoutExtension(path);
+            string prefabPath = Path.Combine(folder, Importer.PREFABS_FOLDER, name + ".prefab");
+            if (File.Exists(prefabPath))
+                return AssetDatabase.LoadAssetAtPath<GameObject>(prefabPath);
+            return null;
+        }        
 
-        public static GameObject GetPrefabFromObject(Object obj)
+        public static GameObject GetSourcePrefabFromObject(Object obj)
         {            
             Object source = PrefabUtility.GetCorrespondingObjectFromSource(obj);
             if (source)
@@ -641,7 +655,7 @@ namespace Reallusion.Import
 
         public static GameObject GetRootPrefabFromObject(GameObject obj)
         {            
-            GameObject prefab = GetPrefabFromObject(obj);
+            GameObject prefab = GetSourcePrefabFromObject(obj);
 
             if (prefab)
             {